requirements.txt
asgiref==3.3.1
click==7.1.2
Django==3.1.7
django-extensions==3.1.1
django-health-check==3.16.3
djangorestframework==3.12.4
panflute==2.0.5
psutil==5.8.0
psycopg2-binary==2.8.6
pypandoc==1.5
python-dotenv==0.16.0
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
